refactor(mllp-client): extract reconnect/backoff to util/; drop dead error state
- Move JitterStrategy / ReconnectPolicy / NO_RECONNECT / DEFAULT_RECONNECT_POLICY
  / backoffDelay out of state.ts into src/util/reconnect.ts — a pure policy +
  backoff-math helper (no state, no I/O). state.ts is now just the machine.
- Remove the machine's `lastError` context field and the `recordError` action:
  nothing ever read `context` (verified), so it was write-only dead state — not
  "only the last error is surfaced", but an error recorded and never used. The
  real error reaches the caller via the connect() throw and the in-flight send
  rejection. Also drop the now-unused `error?` payloads from the CONNECT_FAILED /
  DROP / RECONNECT_FAILED events and the client sites that passed them.

96 tests pass.
